requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.11.1
google-cloud-secret-manager==2.16.0

